When wrapping an ISchema, if the ICompoundProperty passed in is the top compound of the IObject and hence has no compound parent, then use the default schema name to get the child property.
Move init and createSelfBoundsProperty  on OGeomBase to protected, as only
inherited classes need it.

Allow points on OPolyMesh to be part of the sparse set, and initialize the
mandatory properties in init again if we are not doing a sparse set.

Fix another signed/unsigned mismatch in Naming.

+++ lib/Alembic/Util/Naming.cpp
@@ -90,7 +90,7 @@ void makeStandardName( std::string& name, const char fixChar )
         name[0] = fixChar;
     }
 
-    for ( int ii=1; ii<length; ++ii )
+    for ( size_t ii=1; ii<length; ++ii )
     {
         if ( !isStandardChar( name[ii] ) )
         {
